pyproject.toml
[project]
name = "spam-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mlflow>=3.8.1",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "pytorch-lightning>=2.6.0",
    "scikit-learn>=1.8.0",
    "transformers>=4.57.3",
]
